+++ backend/src/main/java/com/jarom/funbankapp/model/User.java
@@ -1,19 +1,19 @@
 package com.jarom.funbankapp.model;
 
 public class User {
-    private int id;
+    private Long id;
     private String username;
     private String email;
     private String password;
 
     // Getters
-    public int getId() { return id; }
+    public Long getId() { return id; }
     public String getUsername() { return username; }
     public String getEmail() { return email; }
     public String getPassword() { return password; }
 
     // Setters
-    public void setId(int id) { this.id = id; }
+    public void setId(Long id) { this.id = id; }
     public void setUsername(String username) { this.username = username; }
     public void setEmail(String email) { this.email = email; }
     public void setPassword(String password) { this.password = password; }
